pyproject.toml
[project]
name = "cdc-spike"
version = "0.1.0"
description = "CDC proof of concept: PostgreSQL → Debezium → Kafka → Elasticsearch"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "kafka-python>=2.0.2",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "rich>=13.9.0",
    "elasticsearch>=8.11.0,<9.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
]

[tool.hatch.build.targets.wheel]
packages = ["src/cdc_spike"]
